requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
discord.py==1.5.1
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
json5==0.9.5
jsonschema==3.2.0
matplotlib==3.2.2
numpy==1.18.5
requests==2.24.0
urllib3==1.25.9